Move files under new alerting home directory
